*Finally* able to get some work done on the church
added modular walls & roofing to the church. It looks ugly right now, but I wanted to get a baseline down, gonna try to mess around with the texturing that way it's not all same-y. Might end up using the thatch roof and replacing the texture with a wooden one, since it's thicker that way, and has a bumpier look to it (since it looks more natural, less perfect that way).